added navigation through a video file. optimization by memory

+++ examples/example_timer.cpp
@@ -4,12 +4,12 @@
 
 #include "mlib/timer.hpp"
 
-void foo(double * arr, int n)
+void foo(double * arr, size_t n)
 {
-	for (int i1 = 0; i1 < n; i1++)
-	for (int i2 = 0; i2 < n; i2++)
-	for (int i3 = 0; i3 < n; i3++)
-	for (int j = 0; j < n; j++)
+	for (size_t i1 = 0; i1 < n; i1++)
+	for (size_t i2 = 0; i2 < n; i2++)
+	for (size_t i3 = 0; i3 < n; i3++)
+	for (size_t  j = 0;  j < n; j++)
 		arr[j] = sin(i1 * i2 * i3 * j);
 }
 
@@ -19,7 +19,7 @@ int main(int, const char **)
 
 	timer.start();
 
-	int size = 64;
+	size_t size = 64;
 	double * arr = new double[size];
 	foo(arr, size);
 	std::cout << "arr[size / 2] = " << arr[size / 2] << std::endl;
